[BUGFIX:PONG] Front-end game crash due to small resize singularity fixed.
A bug caused the game's rendering engine to enter an undefined behaviour loop when the canvas was resized to be extremely small. I am fairly certain this is because at some point the floats to calculate rescales would have been less than float epsilon, converted to 0, and then could no longer be rescaled upwards. The fix is simply a lower limit on canvas size defined as one half the reference resolution.